refactors gameplay into own directory
