+++ gameplay/internal/alts/alter_ego.go
