+++ gameplay/internal/alts/alter_ego_test.go
@@ -3,7 +3,7 @@ package alts_test
 import (
 	"testing"
 
-	"github.com/juggler434/marvelServer/internal/alts"
+	"github.com/juggler434/marvelServer/gameplay/internal/alts"
 )
 
 const (
