+++ gameplay/internal/identity/base_identity_test.go
@@ -3,9 +3,9 @@ package identity_test
 import (
 	"testing"
 
-	"github.com/juggler434/marvelServer/internal/alts"
-	"github.com/juggler434/marvelServer/internal/heroes"
-	"github.com/juggler434/marvelServer/internal/identity"
+	"github.com/juggler434/marvelServer/gameplay/internal/alts"
+	"github.com/juggler434/marvelServer/gameplay/internal/heroes"
+	"github.com/juggler434/marvelServer/gameplay/internal/identity"
 )
 
 const (
