+++ gameplay/main.go
@@ -1,0 +1,7 @@
+package gameplay
+
+import "github.com/juggler434/marvelServer/gameplay/cmd"
+
+func main() {
+	cmd.Execute()
+}
