--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/juggler434/marvelServer/cmd"
-
-func main() {
-	cmd.Execute()
-}
